Support exclude feature for robot and junit checkers
