+++ src/mlx/regex_checker.py
@@ -19,20 +19,6 @@ class RegexChecker(WarningsChecker):
     name = 'regex'
     pattern = None
 
-    def add_patterns(self, regexes, pattern_container):
-        ''' Adds regexes as patterns to the specified container
-
-        Args:
-            regexes (list[str]|None): List of regexes to add
-            pattern_container (list[re.Pattern]): Target storage container for patterns
-        '''
-        if regexes:
-            if not isinstance(regexes, list):
-                raise TypeError("Expected a list value for exclude key in configuration file; got {}"
-                                .format(regexes.__class__.__name__))
-            for regex in regexes:
-                pattern_container.append(re.compile(regex))
-
     def check(self, content):
         ''' Function for counting the number of warnings in a specific text
 
@@ -46,32 +32,6 @@ class RegexChecker(WarningsChecker):
                 continue
             self.count += 1
             self.print_when_verbose(match_string)
-
-    def _is_excluded(self, content):
-        ''' Checks if the specific text must be excluded based on the configured regexes for exclusion and inclusion.
-
-        Inclusion has priority over exclusion.
-
-        Args:
-            content (str): The content to parse
-
-        Returns:
-            bool: True for exclusion, False for inclusion
-        '''
-        matching_exclude_pattern = self._search_patterns(content, self.exclude_patterns)
-        if not self._search_patterns(content, self.include_patterns) and matching_exclude_pattern:
-            self.print_when_verbose("Excluded {!r} because of configured regex {!r}"
-                                    .format(content, matching_exclude_pattern))
-            return True
-        return False
-
-    @staticmethod
-    def _search_patterns(content, patterns):
-        ''' Returns the regex of the first pattern that matches specified content, None if nothing matches '''
-        for pattern in patterns:
-            if pattern.search(content):
-                return pattern.pattern
-        return None
 
 
 class CoverityChecker(RegexChecker):
